requirements.txt
google-cloud-firestore
avro-python3==1.10.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-JSON==0.3.4
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-core==2.0.1
google-auth==2.0.2
google-cloud==0.34.0
google-cloud-pubsub==2.8.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpc-google-iam-v1==0.12.3
grpcio==1.40.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.20
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
python-http-client==3.3.2
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
secure-smtplib==0.1.1
sendgrid==6.8.1
six==1.16.0
SQLAlchemy==1.4.23
starkbank-ecdsa==1.1.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
